greenlight-lib/src/checks/rootfs.rs: Add more debugging info around the rootfs_readonly check
This will help us find out if the reason the check failed is what we
expect. For example, this check was tested both on Universal blue (local
bazzite machine)
environments, bootc, and coreos environments. Though all of them have
immutable bases (ostree or bootc), the universal blue
environment failed the check since it's root filesystem has the
ST_NOATIME instead of ST_RDONLY flag as we would expect. On the other
hand the centos bootc and fedora coreos pass the check as their root
filesystem is truly marked as readonly from the kernels perspective.