greenlight-lib/src/checks/check.rs: Add logic to run the checks
We will have a pattern where we can loop through a set of checks and run
each check as a method with the pattern being:
 check.run()

Currently we only have the "rootfs_readonly" check but will expand when
the other checks have been implemented. For now any other check should
result in an unsupported deployment error.