greenlight-cli/src/main.rs: Add the logic to run checks in the CLI
We might have to think a bit more about how these checks should be
handled. For now the checks follow the following format:
- If it passes all checks it prints a message and returns the `EXITCODE::SUCCESS`
- It exits with an error for any problem conducting the check or if the
  check gracefully fails

At this time we are going with a critical approach where one error means
we should stop any other test, but it could be interesting to have
another mode that would continue the test and only report the error,
maybe with a threshold for how many failed checks until its critical.