README.md: Update the available checks and example config
There's no need for specific checks like `microshift_installed` or
`sshd_enabled` and
such since we can just use the unit_state to check if its up.
eventually it might be a good idea to a another option to unit_state
that ensures the services are up or down. This would be good since when
deploying deploying the DPU Operator and other services to bootstrap the
operator we can ensure its there using this lightweight tool. Then this
on boot time it can act as a lifeline for the operator in case even the
cluster fully disconnects. We could employ the same logic with rhc and
ensure its connection to RHEL Insights in RHEL environments.