Containerfile.microshift: Add test bootc container for microshift
This is a test container that uses rhel-bootc with microshift inside to
build and boot it in a vm with greenlight integrating with greenbooth
health checks. The idea is to have a fast feedback loop by building
this image and pushing to a local registry, and use `bootc upgrade` to
apply atomic upgrades with our new binary. Since we are using a local vm
and local registry these upgrades are super fast and works well as a dev
loop. Of course to get microshift to fully work you must add the pull
secret to the right place and a general guide can be found below:
https://github.com/openshift/microshift/blob/14296c0cfad5db0fd95cca19e092c2cd1ea5727e/docs/contributor/image_mode.md