greenlight-cli/src/main.rs: Fix the way checks are ran
The `--only` flag was not being taken into account when the checks were
ran. The behavior of the wanted vs required checks was also changed such
that wanted checks continue being run but in the end returns an error if
any of them fails. On the other hand a required check failing causes an
immediate error being returned. This behavior allows easy integration
with greenboot.