greenlight-lib/../network.rs: Add logic for validating net interfaces
The plan is for this check is to validate the attributes of the link.
To accomplish this I added an async function that iterates through the
links and matches them based on the name given on the configuration.
It then compares the expected vs actual state of these attributes and
determines if the interface matches the expected one.