greenlight-lib: Add rtnetlink for networking functionality
Adding rtnetlink allows us to interact with networking functionality which seems to be the most robust
way to interact such as managing interfaces without shelling out to
external environments. May be overkill for just
viewing interfaces but it would be a good idea to be able to validate
and ensure virtual functions and other more low level things quickly for specialized
hardware like DPUs.