greenlight-lib/src/checks/unit.rs: Transition from zbus to async systemctl based checks
This is the best of both worlds in terms of complexity and dependency
and should be enough for our use case. If shelling out to systemctl
become costly we can move towards using something like the dbus crate.

Our intent here is to fail only when a proper error happens, and letting
the cli handle how it behaves based on what the importance is set to.