greenlight-lib/.../check.rs: Consume wait service logic with timeout
We will now have a timeout that will be based on the config give for the
unit. It will default to 0 in the config but will be useful for services
like `microshift.service` that takes a bit to come up properly. It will
loop until a timeout is reached or when it finds the expected state.