Containerfile.microshift: Switch to a better base image
We were using a test base image that didn't have the valid entitlements
to install packages from rhsm.
Right now this file is not easily reproduceble, since you need a container
that's registered with RHSM. Please tailor it to your own environment.
Running the build container in a RHEL registered host is highly
recommended.